refactor: structure reorganization
