--- client/src/components/AuthenticationButton/AuthenticationButton.scss
@@ -1,2 +1,0 @@
-@import '../../custom-styles.scss';
-
