--- client/src/components/CreateEntryForm/CreateEntryForm.scss
@@ -1,56 +1,0 @@
-@import '../../custom-styles.scss';
-
-.create-entry-form {
-  form {
-    label {
-      margin-bottom: 5px;
-      font-size: 0.875em;
-      display: block;
-    }
-
-    input {
-      margin-bottom: 10px;
-    }
-    
-    textarea {
-      margin-bottom: 10px;
-    }
-
-    .mantine-TextInput-error {
-      margin-bottom: 15px;
-    }
-
-    .mantine-RichTextEditor-root {
-      margin-bottom: 10px;
-    }
-
-    .mantine-MultiSelect-root {
-      margin-bottom: 10px;
-    }
-
-    .mantine-1ya1dhl .ql-editor {
-      min-height: 200px;
-      height: 100%;
-    }
-
-    .create-entry-form-submit {
-      font-size: 0.875em;
-      font-family: $font-medium;
-      background-color: $color-purple;
-      border: none;
-      border-radius: 8px;
-      line-height: 1em;
-      padding: 13px 16px;
-      color: #fff;
-      margin-right: 10px;
-      margin-bottom: 10px;
-      cursor: pointer;
-      transition: all 0.3s ease;
-
-      &:hover {
-        background-color: darken($color-purple, 5);
-        transition: all 0.3s ease;
-      }
-    }
-  }
-}
