--- client/src/components/EntryList/EntryList.scss
@@ -1,8 +1,0 @@
-@import '../../custom-styles.scss';
-
-.entry-list {
-  h1 {
-    margin-top: 50px;
-    transition: all 0.3s ease;
-  }
-}
