+++ client/src/pages/Home/components/SpacesList/SpacesList.scss
@@ -1,0 +1,71 @@
+@import '../../../../custom-styles';
+
+.space-item-wrapper {
+  position: relative;
+  text-align: center;
+  color: #fff;
+  margin: 10px;
+  min-width: calc(100% / 5 - 20px); // 5 items per row
+  flex-wrap: wrap;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  border-radius: 8px;
+  box-shadow: $shadow-default;
+  transition: all 0.3s ease;
+  cursor: pointer;
+  text-decoration: none;
+  background-color: $color-dark-green;
+
+  &:hover {
+    transform: scale(1.01);
+    transition: all 0.3s ease;
+  }
+
+  .mantine-Menu-root {
+    position: absolute;
+    right: 5px;
+    border-radius: 8px;
+    top: 2px;
+
+    &:hover {
+      background: none !important;
+    }
+
+    button {
+      color: #fff;
+
+      &:hover {
+        background: none !important;
+      }
+    }
+  }
+
+  .spaces-item {
+    height: 100%;
+    width: 100%;
+    padding: 40px 15px;
+    text-decoration: none;
+    color: #fff;
+
+    img {
+      width: 50px;
+      height: 50px;
+      border-radius: 100%;
+      margin-bottom: 15px;
+    }
+
+    .spaces-item-name {
+      font-size: 1em;
+      line-height: 1em;
+      margin-bottom: 10px;
+    }
+
+    .spaces-item-creator {
+      font-size: 0.8em;
+      font-family: $font-regular;
+      line-height: 1em;
+    }
+  }
+}
