+++ client/src/pages/Profile/Profile.scss
@@ -1,0 +1,53 @@
+@import '../../custom-styles';
+
+.profile {
+  margin-top: 80px;
+  text-align: center;
+
+  h1 {
+    line-height: 1em;
+    margin-bottom: 30px;
+  }
+
+  p {
+    width: 400px;
+    margin: 0 auto 20px;
+    text-align: left;
+  }
+
+  .profile-data {
+    width: 400px;
+    margin: 0 auto;
+
+    form {
+      label {
+        display: block;
+        margin-bottom: 10px;
+        text-align: left;
+      }
+      input {
+        display: block;
+        border: 1px solid $color-border-grey;
+        padding: 15px 15px;
+        border-radius: 8px;
+        margin-bottom: 20px;
+        width: 100%;
+      }
+      button {
+        background-color: $color-dark-green;
+        color: #fff;
+        border-radius: 8px;
+        border: none;
+        padding: 15px 20px;
+        width: 100%;
+        cursor: pointer;
+        transition: all 0.3s ease;
+
+        &:hover {
+          background-color: lighten($color-dark-green, 5);
+          transition: all 0.3s ease;
+        }
+      }
+    }
+  }
+}
