+++ client/src/pages/Space/components/EntryList/EntryList.scss
@@ -1,0 +1,8 @@
+@import '../../../../custom-styles';
+
+.entry-list {
+  h1 {
+    margin-top: 50px;
+    transition: all 0.3s ease;
+  }
+}
