+++ client/src/pages/Space/components/EntryListItem/EntryListItem.scss
@@ -1,0 +1,117 @@
+@import '../../../../custom-styles';
+
+.entry {
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-start;
+  cursor: pointer;
+  opacity: 1;
+  transition: all 0.3s ease;
+  visibility: visible;
+
+  &.active {
+    .entry-right {
+      transition: all 0.3s ease;
+      border-left: 2px solid $color-purple;
+    }
+  }
+  
+  &:hover {
+    .entry-right {
+      transition: all 0.3s ease;
+      border-left: 2px solid $color-purple;
+    }
+  }
+  
+  .entry-left {
+    padding: 32px 48px 0 0;
+
+    .entry-left-date {
+      color: $color-purple;
+      line-height: 1em;
+      font-size: 0.875em;
+      width: 60px;
+    }
+  }
+
+  .entry-right {
+    border-left: 2px solid $color-border-grey;
+    padding: 32px 48px 32px 48px;
+    transition: all 0.3s ease;
+
+    .entry-right-title {
+      color: $color-headings;
+      font-size: 1.125em;
+      line-height: 1em;
+      margin-bottom: 8px;
+    }
+
+    .entry-right-text {
+      color: $color-text;
+      line-height: 1.3em;
+      font-size: 0.875em;
+      margin-bottom: 10px;
+    }
+
+    .entry-right-tags {
+      display: flex;
+      justify-content: flex-start;
+      align-items: center;
+      margin-left: -10px; // REMOVE IF BORDERS ON HOVER
+
+      .comment-tag,
+      .attachment-tag {
+        font-size: 0.875em;
+        line-height: 1em;
+        padding: 10px 10px 10px 40px;
+        margin-right: 10px;
+        position: relative;
+        transition: all 0.3s ease;
+        border-radius: 8px;
+        cursor: pointer;
+
+        &::before {
+          content: '';
+          width: 20px;
+          height: 20px;
+          background-position: center;
+          background-repeat: no-repeat;
+          background-size: contain;
+          position: absolute;
+          left: 10px;
+          top: 50%;
+          transform: translateY(-50%);
+        }
+      }
+
+      .comment-tag {
+        color: $color-purple;
+
+        // &:hover {
+        //   border: 1px solid $color-purple;
+        //   transition: all 0.3s ease;
+        // }
+
+        &::before {
+          content: '';
+          background-image: url(../../../../assets/img/icon-comment.svg);
+        }
+      }
+
+      .attachment-tag {
+        color: $color-light-green;
+        padding-left: 38px;
+
+        // &:hover {
+        //   border: 1px solid $color-light-green;
+        //   transition: all 0.3s ease;
+        // }
+
+        &::before {
+          content: '';
+          background-image: url(../../../../assets/img/icon-attachment.svg);
+        }
+      }
+    }
+  }
+}
